KSLAB-1655: switch course fixed credit options to use float not int
git-svn-id: https://svn.kuali.org/repos/student/trunk@18106 0ac02d80-fb3f-11dc-a574-c57244a850eb

+++ ks-lum/ks-lum-impl/src/main/java/org/kuali/student/lum/course/service/assembler/CourseAssembler.java
@@ -565,7 +565,7 @@ public class CourseAssembler implements BOAssembler<CourseInfo, CluInfo> {
 					Map<String,String> attributes = null;
 					//Depending on the type, set the id, type and result values differently
 					if(CourseAssemblerConstants.COURSE_RESULT_COMP_TYPE_CREDIT_FIXED.equals(creditOption.getType())){
-						int fixedCreditValue = Integer.parseInt(creditOption.getAttributes().get(CourseAssemblerConstants.COURSE_RESULT_COMP_ATTR_FIXED_CREDIT_VALUE));
+						float fixedCreditValue = Float.parseFloat(creditOption.getAttributes().get(CourseAssemblerConstants.COURSE_RESULT_COMP_ATTR_FIXED_CREDIT_VALUE));
 						id = CourseAssemblerConstants.COURSE_RESULT_COMP_CREDIT_PREFIX + fixedCreditValue;
 						type = CourseAssemblerConstants.COURSE_RESULT_COMP_TYPE_CREDIT_FIXED;
 						resultValues = new ArrayList<String>();
